выделяет отрисовку точек путешествия в отдельный пресентор
